+++ docs/src/layouts/MainLayout.astro
@@ -144,6 +144,7 @@ const githubEditUrl = `${CONFIG.GITHUB_EDIT_URL}/${currentFile}`;
     </style>
     <script>
       import 'media-chrome';
+      import 'media-chrome/dist/menu/index';
       import 'media-chrome/dist/media-theme-element';
       import 'media-chrome/dist/experimental/index';
     </script>
