+++ examples/vanilla/control-elements/media-audio-track-menu.html
@@ -5,6 +5,7 @@
     <title>Media Chrome Audio Track Menu</title>
     <script type="module" src="https://cdn.jsdelivr.net/npm/@mux/mux-video@canary/+esm"></script>
     <script type="module" src="../../../dist/index.js"></script>
+    <script type="module" src="../../../dist/menu/index.js"></script>
     <style>
       /** add styles to prevent CLS (Cumulative Layout Shift) */
       media-controller:not([audio]) {
