+++ examples/vanilla/control-elements/media-playback-rate-menu.html
@@ -4,6 +4,7 @@
     <meta name="viewport" content="width=device-width">
     <title>Media Chrome Playback Rate Menu</title>
     <script type="module" src="../../../dist/index.js"></script>
+    <script type="module" src="../../../dist/menu/index.js"></script>
     <style>
       /* Hide custom elements that are not defined yet */
       :not(:defined) {
