+++ examples/vanilla/control-elements/media-settings-menu.html
@@ -5,6 +5,7 @@
     <title>Media Settings Menu</title>
     <script type="module" src="https://cdn.jsdelivr.net/npm/@mux/mux-video/+esm"></script>
     <script type="module" src="../../../dist/index.js"></script>
+    <script type="module" src="../../../dist/menu/index.js"></script>
     <style>
       /* Hide custom elements that are not defined yet */
       :not(:defined) {
