+++ src/js/index.ts
@@ -6,16 +6,10 @@ export * as timeUtils from './utils/time.js';
 // when calling `associateElement(this)` in connectedCallback.
 import MediaController from './media-controller.js';
 import MediaAirplayButton from './media-airplay-button.js';
-import MediaAudioTrackMenu from './media-audio-track-menu.js';
-import MediaAudioTrackMenuButton from './media-audio-track-menu-button.js';
 import MediaCaptionsButton from './media-captions-button.js';
-import MediaCaptionsMenu from './media-captions-menu.js';
-import MediaCaptionsMenuButton from './media-captions-menu-button.js';
 import MediaCastButton from './media-cast-button.js';
 import MediaChromeButton from './media-chrome-button.js';
 import MediaChromeDialog from './media-chrome-dialog.js';
-import MediaChromeMenu from './media-chrome-menu.js';
-import MediaChromeMenuItem from './media-chrome-menu-item.js';
 import MediaChromeRange from './media-chrome-range.js';
 import MediaControlBar from './media-control-bar.js';
 import MediaDurationDisplay from './media-duration-display.js';
@@ -26,20 +20,13 @@ import MediaLoadingIndicator from './media-loading-indicator.js';
 import MediaMuteButton from './media-mute-button.js';
 import MediaPipButton from './media-pip-button.js';
 import MediaPlaybackRateButton from './media-playback-rate-button.js';
-import MediaPlaybackRateMenu from './media-playback-rate-menu.js';
-import MediaPlaybackRateMenuButton from './media-playback-rate-menu-button.js';
 import MediaPlayButton from './media-play-button.js';
 import MediaPosterImage from './media-poster-image.js';
 import MediaPreviewChapterDisplay from './media-preview-chapter-display.js';
 import MediaPreviewThumbnail from './media-preview-thumbnail.js';
 import MediaPreviewTimeDisplay from './media-preview-time-display.js';
-import MediaRenditionMenu from './media-rendition-menu.js';
-import MediaRenditionMenuButton from './media-rendition-menu-button.js';
 import MediaSeekBackwardButton from './media-seek-backward-button.js';
 import MediaSeekForwardButton from './media-seek-forward-button.js';
-import MediaSettingsMenu from './media-settings-menu.js';
-import MediaSettingsMenuButton from './media-settings-menu-button.js';
-import MediaSettingsMenuItem from './media-settings-menu-item.js';
 import MediaTimeDisplay from './media-time-display.js';
 import MediaTimeRange from './media-time-range.js';
 import MediaTooltip from './media-tooltip.js';
@@ -47,16 +34,10 @@ import MediaVolumeRange from './media-volume-range.js';
 
 export {
   MediaAirplayButton,
-  MediaAudioTrackMenu,
-  MediaAudioTrackMenuButton,
   MediaCaptionsButton,
-  MediaCaptionsMenu,
-  MediaCaptionsMenuButton,
   MediaCastButton,
   MediaChromeButton,
   MediaChromeDialog,
-  MediaChromeMenu,
-  MediaChromeMenuItem,
   MediaChromeRange,
   MediaControlBar,
   MediaController,
@@ -68,20 +49,13 @@ export {
   MediaMuteButton,
   MediaPipButton,
   MediaPlaybackRateButton,
-  MediaPlaybackRateMenu,
-  MediaPlaybackRateMenuButton,
   MediaPlayButton,
   MediaPosterImage,
   MediaPreviewChapterDisplay,
   MediaPreviewThumbnail,
   MediaPreviewTimeDisplay,
-  MediaRenditionMenu,
-  MediaRenditionMenuButton,
   MediaSeekBackwardButton,
   MediaSeekForwardButton,
-  MediaSettingsMenu,
-  MediaSettingsMenuButton,
-  MediaSettingsMenuItem,
   MediaTimeDisplay,
   MediaTimeRange,
   MediaTooltip,
