+++ src/js/menu/media-playback-rate-menu-button.ts
@@ -1,13 +1,13 @@
-import { globalThis, document } from './utils/server-safe-globals.js';
-import { MediaUIAttributes } from './constants.js';
-import { nouns, tooltipLabels } from './labels/labels.js';
+import { globalThis, document } from '../utils/server-safe-globals.js';
+import { MediaUIAttributes } from '../constants.js';
+import { nouns, tooltipLabels } from '../labels/labels.js';
 import { MediaChromeMenuButton } from './media-chrome-menu-button.js';
-import { AttributeTokenList } from './utils/attribute-token-list.js';
+import { AttributeTokenList } from '../utils/attribute-token-list.js';
 import {
   getNumericAttr,
   setNumericAttr,
   getMediaController,
-} from './utils/element-utils.js';
+} from '../utils/element-utils.js';
 
 export const Attributes = {
   RATES: 'rates',
