+++ src/js/menu/media-settings-menu-item.ts
@@ -1,4 +1,4 @@
-import { globalThis, document } from './utils/server-safe-globals.js';
+import { globalThis, document } from '../utils/server-safe-globals.js';
 import { MediaChromeMenuItem } from './media-chrome-menu-item.js';
 
 const template: HTMLTemplateElement = document.createElement('template');
